TASK: Move introduced `SoftRemoval` into new `SubtreeTagging` namespace
