+++ Neos.Neos/Classes/Domain/SubtreeTagging/SoftRemoval/SoftRemovalConstraintsCommandHookFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Neos\Neos\Domain\SoftRemoval;
+namespace Neos\Neos\Domain\SubtreeTagging\SoftRemoval;
 
 use Neos\ContentRepository\Core\CommandHandler\CommandHookInterface;
 use Neos\ContentRepository\Core\Factory\CommandHookFactoryInterface;
