+++ Neos.Neos/Configuration/Settings.ContentRepositoryRegistry.yaml
@@ -17,7 +17,7 @@ Neos:
             'Neos.Neos:AssetUsage':
               factoryObjectName: Neos\Neos\AssetUsage\CatchUpHook\AssetUsageCatchUpHookFactory
             'Neos.Neos:ImpendingHardRemovalConflictDetection':
-              factoryObjectName: Neos\Neos\Domain\SoftRemoval\ImpendingHardRemovalConflictDetectionHookFactory
+              factoryObjectName: Neos\Neos\Domain\SubtreeTagging\SoftRemoval\ImpendingHardRemovalConflictDetectionHookFactory
 
         projections:
           'Neos.Neos:DocumentUriPathProjection':
